+++ client/components/src/index.js
@@ -1,3 +1,3 @@
 export { default as Sheaf } from "./Sheaf.svelte"
 export { default as IntersectionPoint } from "./IntersectionPoint.svelte"
-export { default as Tabview } from "./stubs/Tabview.svelte"
+export { default as Tabview } from "./Tabview.svelte"
